airbyte-integrations/connectors/source-endpoint-central/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-endpoint-central"
description = "Source implementation for endpoint-central."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/endpoint-central"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_endpoint_central" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
airbyte-cdk = "^0"

[tool.poetry.scripts]
source-endpoint-central = "source_endpoint_central.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"

